refactor: extract formatter and handler classes from logging into _log_format
Moves ColorFormatter, StderrHandler, and color detection into
_log_format.py. logging.py keeps logger setup and log_* helpers.
